Application,Mail Service and FileRepo

+++ backend/Hr_Portalgrad/Services/UserReporsitories/DbUserReporsitory.cs
@@ -44,11 +44,16 @@ namespace HR_Portalgrad.Services.UserReporsitories
             return await _context.Users.ToListAsync();
         }
 
-        public async Task<User> UpdatetoHr(int id,string userRole)
+        public async Task<User> UpdatetoHr(int id, string Passwords, string firstName, string lastName, string userRole, string email)
         {
 
-            var us = await _context.Users.FindAsync(id);
+            var us = new User { Id = id, Passwords = Passwords, firstName = firstName, lastName = lastName, userRole = userRole, email = email };
             _context.Users.Attach(us);
+            _context.Entry(us).Property(j => j.Id).IsModified = false;
+            _context.Entry(us).Property(j => j.firstName).IsModified = false;
+            _context.Entry(us).Property(j => j.lastName).IsModified = false;
+            _context.Entry(us).Property(j => j.email).IsModified = false;
+            _context.Entry(us).Property(x => x.Passwords).IsModified = false;
             _context.Entry(us).Property(x => x.userRole).IsModified = true;
             await _context.SaveChangesAsync();
             return us;
@@ -58,14 +63,18 @@ namespace HR_Portalgrad.Services.UserReporsitories
             var us=await _context.Users.FindAsync(id);
             var us2 =await _context.Applicants.FirstOrDefaultAsync(u=>u.UserId==id);
            _context.Users.Remove(us);
-            _context.Applicants.Remove(us2);
+            if (us2 != null)
+            {
+                _context.Applicants.Remove(us2);
+            }
+            
             await _context.SaveChangesAsync();
             return us;
 
         }
         public async Task<IEnumerable> GetApplicantsbyUser()
         {
-
+            
             var us = await (from p in _context.Users
                             join p2 in _context.Applicants
                             on p.Id equals p2.UserId
@@ -81,18 +90,29 @@ namespace HR_Portalgrad.Services.UserReporsitories
                                 p2.UserId,
                                 p2.ProfileId
                             }).ToListAsync();
-            //var us2 = new Applicants { UserId=id, isAccepted = isAccepted };
-            //var us3 = await _context.Jobs.Where(u=>u.UserId == id).FirstOrDefaultAsync();
-            //var us = await _context.Users.FindAsync(id);
-
-
-            //await _context.Applicants.Where(u => u.UserId == id).SingleOrDefaultAsync();
-            //_context.Applicants.Attach(us2);
-            //_context.Entry(us2).Property(x => x.isAccepted).IsModified = true;
+            
 
             await _context.SaveChangesAsync();
             return us;
         }   
+        public async Task<IEnumerable> GetUserProfile()
+        {
+            var us = await (from p in _context.Users
+                            join p2 in _context.profiles
+                            on p.Id equals p2.Userid
+                            select new
+                            {
+                                p.firstName,
+                                p.lastName,
+                                p.phoneNumber,
+                                p.country,
+                                p.gender,
+                                p2.Id,
+                                p2.Userid
+                            }).ToListAsync();
+            await _context.SaveChangesAsync();
+            return us;
+        }
     
         public async Task<User> ChangePassword(int id, string Passwords,string firstName,string lastName,string userRole,string email)
         {
@@ -108,6 +128,74 @@ namespace HR_Portalgrad.Services.UserReporsitories
             await _context.SaveChangesAsync();
             return us;
         }
-      
+
+        public async Task<User> EditUserCountry(int id, string firstName, string lastName, string email, string Passwords, string userRole, string country)
+        {
+            var user = new User { Id = id, firstName = firstName, lastName = lastName, email = email, Passwords = Passwords, userRole = userRole, country = country };
+            _context.Users.Attach(user);
+            _context.Entry(user).Property(j => j.Id).IsModified = false;
+         
+            _context.Entry(user).Property(j => j.firstName).IsModified = false;
+            _context.Entry(user).Property(j => j.lastName).IsModified = false;
+            _context.Entry(user).Property(j => j.email).IsModified = false;
+            _context.Entry(user).Property(j => j.Passwords).IsModified = false;
+            _context.Entry(user).Property(j => j.userRole).IsModified = false;
+            _context.Entry(user).Property(j => j.country).IsModified = true;
+            await _context.SaveChangesAsync();
+            return user;
+        }
+
+        public async Task<User> EditUserPhone(int id, string phone)
+        {
+            var user = new User { Id = id, phoneNumber = phone };
+            _context.Users.Attach(user);
+            _context.Entry(user).Property(j => j.Id).IsModified = false;
+            _context.Entry(user).Property(j => j.phoneNumber).IsModified = true;
+            await _context.SaveChangesAsync();
+            return user;
+        }
+
+        public async Task<User> EditUserGender(int id,string firstName,string lastName,string email,string Passwords,string userRole, string gender)
+        {
+            var user = new User { Id = id,firstName=firstName,lastName=lastName,email=email,Passwords=Passwords,userRole=userRole, gender = gender };
+            _context.Users.Attach(user);
+            _context.Entry(user).Property(j => j.Id).IsModified = false;
+            _context.Entry(user).Property(j => j.firstName).IsModified = false;
+            _context.Entry(user).Property(j => j.lastName).IsModified = false;
+            _context.Entry(user).Property(j => j.email).IsModified = false;
+            _context.Entry(user).Property(j => j.Passwords).IsModified = false;
+            _context.Entry(user).Property(j => j.userRole).IsModified = false;
+            _context.Entry(user).Property(j => j.gender).IsModified = true;
+            await _context.SaveChangesAsync();
+            return user;
+        }
+        public async Task<User> EditUserPhone(int id, string firstName, string lastName, string email, string Passwords, string userRole, string phoneNumber)
+        {
+            var user = new User { Id = id, firstName = firstName, lastName = lastName, email = email, Passwords = Passwords, userRole = userRole, phoneNumber = phoneNumber };
+            _context.Users.Attach(user);
+            _context.Entry(user).Property(j => j.Id).IsModified = false;
+            _context.Entry(user).Property(j => j.firstName).IsModified = false;
+            _context.Entry(user).Property(j => j.lastName).IsModified = false;
+            _context.Entry(user).Property(j => j.email).IsModified = false;
+            _context.Entry(user).Property(j => j.Passwords).IsModified = false;
+            _context.Entry(user).Property(j => j.userRole).IsModified = false;
+            _context.Entry(user).Property(j => j.phoneNumber).IsModified = true;
+            await _context.SaveChangesAsync();
+            return user;
+        }
+        public async Task<User> EditEmailValid(int id, string firstName, string lastName, string email, string Passwords, string userRole, int isEmailValid)
+        {
+            var user = new User { Id = id, firstName = firstName, lastName = lastName, email = email, Passwords = Passwords, userRole = userRole, isEmailValid = 1 };
+            _context.Users.Attach(user);
+            _context.Entry(user).Property(j => j.Id).IsModified = false;
+            _context.Entry(user).Property(j => j.firstName).IsModified = false;
+            _context.Entry(user).Property(j => j.lastName).IsModified = false;
+            _context.Entry(user).Property(j => j.email).IsModified = false;
+            _context.Entry(user).Property(j => j.Passwords).IsModified = false;
+            _context.Entry(user).Property(j => j.userRole).IsModified = false;
+            _context.Entry(user).Property(j => j.isEmailValid).IsModified = true;
+            await _context.SaveChangesAsync();
+            return user;
+        }
     }
 }
